Subthought: Re-calculate shim height on cursor change.
